refactor: create new submodule native-library-common to merge NativeUtils
