+++ circe-checksum/pom.xml
@@ -39,6 +39,11 @@
   </properties>
 
   <dependencies>
+    <dependency>
+      <groupId>org.apache.bookkeeper</groupId>
+      <artifactId>native-library-common</artifactId>
+      <version>${project.parent.version}</version>
+    </dependency>
 
     <dependency>
       <groupId>com.google.guava</groupId>
