+++ cpu-affinity/pom.xml
@@ -34,6 +34,11 @@
 
   <dependencies>
     <dependency>
+      <groupId>org.apache.bookkeeper</groupId>
+      <artifactId>native-library-common</artifactId>
+      <version>${project.parent.version}</version>
+    </dependency>
+    <dependency>
       <groupId>com.google.guava</groupId>
       <artifactId>guava</artifactId>
     </dependency>
